+++ rusoto/credential/src/instance_metadata.rs
@@ -4,7 +4,7 @@ use async_trait::async_trait;
 use hyper::Uri;
 use std::time::Duration;
 
-use crate::request::HttpClient;
+use crate::request::{HttpClient, HttpClientFuture};
 use crate::{
     parse_credentials_from_aws_service, AwsCredentials, CredentialsError, ProvideAwsCredentials,
 };
@@ -17,7 +17,7 @@ const AWS_CREDENTIALS_PROVIDER_PATH: &str = "latest/meta-data/iam/security-crede
 /// The provider has a default timeout of 30 seconds. While it should work well for most setups,
 /// you can change the timeout using the `set_timeout` method.
 ///
-/// # Example
+/// # Examples
 ///
 /// ```rust
 /// extern crate rusoto_credential;
@@ -34,10 +34,29 @@ const AWS_CREDENTIALS_PROVIDER_PATH: &str = "latest/meta-data/iam/security-crede
 ///   // ...
 /// }
 /// ```
+///
+/// The source location can be changed from the default of 169.254.169.254:
+///
+/// ```rust
+/// extern crate rusoto_credential;
+///
+/// use std::time::Duration;
+///
+/// use rusoto_credential::InstanceMetadataProvider;
+///
+/// fn main() {
+///   let mut provider = InstanceMetadataProvider::new();
+///   // you can overwrite the default endpoint like this:
+///   provider.set_ip_addr_with_port("127.0.0.1", "8080");
+///
+///   // ...
+/// }
+/// ```
 #[derive(Clone, Debug)]
 pub struct InstanceMetadataProvider {
     client: HttpClient,
     timeout: Duration,
+    metadata_ip_addr: String,
 }
 
 impl InstanceMetadataProvider {
@@ -46,12 +65,18 @@ impl InstanceMetadataProvider {
         InstanceMetadataProvider {
             client: HttpClient::new(),
             timeout: Duration::from_secs(30),
+            metadata_ip_addr: AWS_CREDENTIALS_PROVIDER_IP.to_string(),
         }
     }
 
     /// Set the timeout on the provider to the specified duration.
     pub fn set_timeout(&mut self, timeout: Duration) {
         self.timeout = timeout;
+    }
+
+    /// Allow overriding host and port of instance metadata service.
+    pub fn set_ip_addr_with_port(&mut self, ip: &str, port: &str) {
+        self.metadata_ip_addr = format!("{}:{}", ip, port);
     }
 }
 
@@ -100,4 +125,39 @@ impl ProvideAwsCredentials for InstanceMetadataProvider {
                 })?;
         parse_credentials_from_aws_service(&cred_str)
     }
+}
+
+/// Gets the role name to get credentials for using the IAM Metadata Service (169.254.169.254).
+fn get_role_name(
+    client: &HttpClient,
+    timeout: Duration,
+    ip_addr: &str,
+) -> Result<HttpClientFuture, CredentialsError> {
+    let role_name_address = format!("http://{}/{}/", ip_addr, AWS_CREDENTIALS_PROVIDER_PATH);
+    let uri = match role_name_address.parse::<Uri>() {
+        Ok(u) => u,
+        Err(e) => return Err(CredentialsError::new(e)),
+    };
+
+    Ok(client.get(uri, timeout))
+}
+
+/// Gets the credentials for an EC2 Instances IAM Role.
+fn get_credentials_from_role(
+    client: &HttpClient,
+    timeout: Duration,
+    role_name: &str,
+    ip_addr: &str,
+) -> Result<HttpClientFuture, CredentialsError> {
+    let credentials_provider_url = format!(
+        "http://{}/{}/{}",
+        ip_addr, AWS_CREDENTIALS_PROVIDER_PATH, role_name
+    );
+
+    let uri = match credentials_provider_url.parse::<Uri>() {
+        Ok(u) => u,
+        Err(e) => return Err(CredentialsError::new(e)),
+    };
+
+    Ok(client.get(uri, timeout))
 }