+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 (Please put changes here)
+- Updated AWS-CREDENTIALS.md to address https://github.com/rusoto/rusoto/issues/1597
 - Fixed SNS API's attributes and value keyword https://github.com/rusoto/rusoto/pull/1591
 - Adding support for web identity provider, which enables IAM roles for Kubernetes service accounts.
 - Add object-safe AwsCredentialsProvider trait as alternative to the existing generic `ProvideAwsCredentials
@@ -14,6 +15,7 @@
 - Encode request payload optionally with Gzip : https://github.com/rusoto/rusoto/pull/1615
 - Add Debug trait to generated Clients
 - Add `rusoto_ec2::filter!` macro
+- Improve `InstanceMetadataProvider` to avoid cloning unnecessarily
 
 ## [0.42.0] - 2019-11-18